+++ authentication_service/decorators.py
@@ -1,4 +1,5 @@
 import warnings
+from functools import wraps
 
 
 class generic_deprecation(object):
@@ -12,3 +13,25 @@ class generic_deprecation(object):
             warnings.warn(self.message, self.warning_class, self.stack_level)
             return method(*args, **kwargs)
         return wrapped
+
+
+def required_form_fields_label_alter(init_func, required_character=None):
+    """
+    Decorator for Django form's __init__ method.
+
+    Appends an asterisk or provided character to labels of fields that are
+    required.
+    """
+    @wraps(init_func)
+    def _wrapped(*args, **kwargs):
+        # Form needs to be initialised before labels are changed
+        init = init_func(*args, **kwargs)
+
+        # Append required character to the existing label
+        for name, field in args[0].fields.items():
+            # NOTE: If the label was not specified on the field, there is
+            # nothing to append to.
+            if field.required and field.label:
+                field.label += f" {required_character or '*'}"
+        return init
+    return _wrapped