+++ authentication_service/request_client/urls.py
@@ -1,0 +1,13 @@
+from django.conf.urls import url
+from django.contrib.auth.decorators import login_required
+
+from authentication_service.request_client import views
+
+
+urlpatterns = [
+    url(
+        r"^request-form/$",
+        login_required(views.RequestCientView.as_view()),
+        name="request_form"
+    )
+]