+++ project/settings.py
@@ -13,11 +13,18 @@ import user_data_store
 
 env = Env()
 
-# Project Settings
+#--Project settings
 HIDE_FIELDS = {
     "global_enable": True,
     "global_fields": ["email", "birth_date", "nickname", "avatar"]
 }
+
+# NOTE: If the value is not set, the celery task will error.
+CLIENT_REQUEST_EMAIL = env.list("CLIENT_REQUEST_EMAIL", [])
+
+# authentication_service/management/commands/user_login_report.py
+REPORT_EMAILS = env.list("REPORT_EMAILS", [])
+#--Project settings end
 
 # Django Settings
 SECRET_KEY = env.str("SECRET_KEY", "_n(_w(3!i4-p((jz8(o0fb*_r5fb5t!qh1g@m9%4vryx5lale=")
@@ -167,7 +174,8 @@ ADDITIONAL_APPS = [
 # Project app has to be first in the list.
 INSTALLED_APPS = [
     "authentication_service",
-    "authentication_service.user_migration"
+    "authentication_service.user_migration",
+    "authentication_service.request_client"
 ] + INSTALLED_APPS + ADDITIONAL_APPS
 
 MIDDLEWARE = [